+++ packages/react/src/components/IonRoute.tsx
@@ -11,11 +11,10 @@ export interface IonRouteProps {
   disableIonPageManagement?: boolean;
 }
 
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface IonRouteState {}
 
 export class IonRoute extends React.PureComponent<IonRouteProps, IonRouteState> {
-  context!: React.ContextType<typeof NavContext>;
+  declare context: React.ContextType<typeof NavContext>;
 
   render() {
     const IonRouteInner = this.context.getIonRoute();