+++ packages/react/src/components/__tests__/utils.spec.ts
@@ -1,5 +1,5 @@
 import * as utils from '../react-component-lib/utils';
-import '@testing-library/jest-dom/extend-expect';
+import '@testing-library/jest-dom';
 
 describe('isCoveredByReact', () => {
   it('should identify standard events as covered by React', () => {
@@ -49,5 +49,43 @@ describe('attachProps', () => {
     expect((div as any).testprop).toEqual(['red']);
     expect(div).toHaveStyle(`display: block;`);
     expect(Object.keys((div as any).__events)).toEqual(['ionClick']);
+  });
+});
+
+describe('attachProps boolean attributes', () => {
+  it('should strip a stray disabled="false" attribute when the prop is false', () => {
+    const div = document.createElement('div');
+    div.setAttribute('disabled', 'false');
+
+    utils.attachProps(div, { disabled: false });
+
+    expect(div.hasAttribute('disabled')).toBe(false);
+  });
+
+  it('should preserve aria-* attributes set to false', () => {
+    const div = document.createElement('div');
+    div.setAttribute('aria-expanded', 'false');
+
+    utils.attachProps(div, { 'aria-expanded': false });
+
+    expect(div.getAttribute('aria-expanded')).toBe('false');
+  });
+
+  it('should preserve data-* attributes set to false', () => {
+    const div = document.createElement('div');
+    div.setAttribute('data-active', 'false');
+
+    utils.attachProps(div, { 'data-active': false });
+
+    expect(div.getAttribute('data-active')).toBe('false');
+  });
+
+  it('should preserve enumerated attributes set to false (e.g. draggable)', () => {
+    const div = document.createElement('div');
+    div.setAttribute('draggable', 'false');
+
+    utils.attachProps(div, { draggable: false });
+
+    expect(div.getAttribute('draggable')).toBe('false');
   });
 });